Moved component to rich editor folder
